move macro tests to macro mod
